+++ R/data.R
@@ -44,6 +44,9 @@
 #' }
 #'
 #' @name signature
+#'
+#' @examples
+#' SBS96.sigs <- signature$genome$SBS96
 "signature"
 
 
@@ -85,6 +88,9 @@
 #' }
 #'
 #' @name exposure
+#'
+#' @examples
+#' SBS96.exposure <- exposure$PCAWG$SBS96
 "exposure"
 
 
@@ -125,6 +131,9 @@
 #' }
 #'
 #' @name spectra
+#'
+#' @examples
+#' SBS96.spectra <- spectra$PCAWG$SBS96
 "spectra"
 
 #' Exposure statistics from the PCAWG7 paper