+++ tests/testthat/test_SampleIDToCancerType.R
@@ -1,0 +1,6 @@
+context("SampleIDToCancerType")
+
+test_that("SampleIDToCancerType", {
+  cancer.type <- SampleIDToCancerType("Biliary-AdenoCA::SP117655")
+  expect_equivalent(cancer.type, "Biliary-AdenoCA")
+})